requirements.txt
asgiref==3.5.2
Django~=4.1.0
djangorestframework==3.13.1
channels==3.0.5
channels-redis==3.4.1
psycopg2-binary==2.9.3
daphne==3.0.2
sqlparse==0.4.2
Pillow==9.2.0
flower==1.2.0
celery[redis]==5.2.7
SQLAlchemy==2.0.2

# for visual studio code
autopep8
pylint
pylint-django
pylint-plugin-utils
flake8
pyflakes
